refactor: move useSwitch to its own folder.
