+++ docs/data/base/components/switch/UseSwitchesBasic.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import clsx from 'clsx';
 import { styled } from '@mui/system';
-import { useSwitch } from '@mui/base/SwitchUnstyled';
+import useSwitch from '@mui/base/useSwitch';
 
 const blue = {
   500: '#007FFF',
