+++ docs/data/base/getting-started/customization/StylingHooks.js
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import clsx from 'clsx';
-import { useSwitch } from '@mui/base/SwitchUnstyled';
+import useSwitch from '@mui/base/useSwitch';
 import { styled } from '@mui/system';
 
 const BasicSwitchRoot = styled('span')`
