+++ packages/mui-base/src/SwitchUnstyled/SwitchUnstyled.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import PropTypes from 'prop-types';
 import { OverridableComponent } from '@mui/types';
 import composeClasses from '../composeClasses';
-import useSwitch from './useSwitch';
+import useSwitch from '@mui/base/useSwitch';
 import { getSwitchUnstyledUtilityClass } from './switchUnstyledClasses';
 import {
   SwitchUnstyledProps,
