+++ packages/mui-base/src/SwitchUnstyled/index.ts
@@ -1,8 +1,5 @@
 export { default } from './SwitchUnstyled';
 export * from './SwitchUnstyled.types';
 
-export { default as useSwitch } from './useSwitch';
-export * from './useSwitch.types';
-
 export { default as switchUnstyledClasses } from './switchUnstyledClasses';
 export * from './switchUnstyledClasses';
