+++ packages/mui-base/src/useSwitch/index.ts
@@ -1,0 +1,5 @@
+export { default as useSwitch } from './useSwitch';
+export { default } from './useSwitch';
+export * from './useSwitch';
+
+export * from './useSwitch.types';
